requirements.txt
altgraph==0.17.2
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
boto3==1.34.69
botocore==1.34.69
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
exceptiongroup==1.2.0
fastapi==0.110.0
flatbuffers==24.3.25
future==0.18.2
h11==0.14.0
httptools==0.6.1
humanfriendly==10.0
idna==3.6
jmespath==1.0.1
macholib==1.15.2
mpmath==1.3.0
NSFWDetection==1.0.2
numpy==1.26.4
onnxruntime==1.17.1
opencv-python-headless==4.9.0.80
packaging==24.0
pillow==10.2.0
progressbar2==4.4.2
protobuf==5.26.0
pydantic==2.6.4
pydantic_core==2.16.3
pydload==1.0.9
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-utils==3.8.2
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
s3transfer==0.10.1
six==1.15.0
sniffio==1.3.1
starlette==0.36.3
sympy==1.12
typing_extensions==4.10.0
urllib3==1.26.18
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
